requirements.dev.txt
# dev dependencies
ipdb
ipython
pudb
pytest-watch
